streamlit_ui/requirements.txt
# Streamlit UI Framework Requirements
# Core dependencies for running Streamlit applications built with the UI Framework

# Core Streamlit
streamlit>=1.28.0

# Environment management
python-dotenv>=1.0.0

# Optional: For enhanced functionality
pandas>=2.0.0          # Data handling
numpy>=1.24.0          # Numerical computing
